Refactor: 패키지 위치 변경 및 ResultDto에 어노테이션 추가
