+++ src/main/java/com/ujutechnology/api8/api/dto/creditLoan/CreditLoanOuterWrapperResult.java
@@ -1,4 +1,4 @@
-package com.ujutechnology.api8.biz.dto.rentHouseLoan;
+package com.ujutechnology.api8.api.dto.creditLoan;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
@@ -11,14 +11,14 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class RentHouseOuterWrapperResult {
+public class CreditLoanOuterWrapperResult {
         private String err_cd; // 응답코드
         private String err_msg; // 응답메시지
         private String total_count; // 총 상품건수
         private String max_page_no; // 총 페이지 건수
         private String now_page_no; // 현재 조회 페이지 번호
 
-        private List<RentHouseBaseList> baseList;
-        private List<RentHouseOptionList> optionList;
+        private List<CreditLoanBaseList> baseList;
+        private List<CreditLoanOptionList> optionList;
 
 }
