+++ src/main/java/com/ujutechnology/api8/api/dto/creditLoan/CreditLoanResult.java
@@ -1,0 +1,12 @@
+package com.ujutechnology.api8.api.dto.creditLoan;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreditLoanResult {
+    CreditLoanOuterWrapperResult result;
+}
