+++ src/main/java/com/ujutechnology/api8/api/dto/deposit/DepositResult.java
@@ -1,0 +1,12 @@
+package com.ujutechnology.api8.api.dto.deposit;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class DepositResult {
+    DepositOuterWrapperResult result;
+}
