+++ src/main/java/com/ujutechnology/api8/api/dto/mortgageLoan/MortgageBaseList.java
@@ -1,4 +1,4 @@
-package com.ujutechnology.api8.biz.dto.creditLoan;
+package com.ujutechnology.api8.api.dto.mortgageLoan;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
@@ -9,14 +9,15 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class CreditLoanBaseList {
+public class MortgageBaseList {
     private String dcls_month; // 공시 제출월
     private String fin_co_no; // 금융회사 코드
     private String kor_co_nm; // 금융회사명
     private String fin_prdt_cd; // 금융상품 코드
     private String fin_prdt_nm; // 금융 상품명
     private String join_way; // 가입 방법
-    private String cb_name; // CB 회사명
-    private String crdt_prdt_type_nm; // 대출종류명
-    private String crdt_prdt_type; // 대출종류 코드
+    private String loan_inci_expn; // 대출 부대비용
+    private String erly_rpay_fee; // 중도 상환 수수료
+    private String dly_rate; // 연체 이자율
+    private String loan_lmt; // 대출한도
 }
