+++ src/main/java/com/ujutechnology/api8/api/dto/mortgageLoan/MortgageResult.java
@@ -1,0 +1,12 @@
+package com.ujutechnology.api8.api.dto.mortgageLoan;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class MortgageResult {
+    MortgageOuterWrapperResult result;
+}
