+++ src/main/java/com/ujutechnology/api8/api/dto/rentHouseLoan/RentHouseResult.java
@@ -1,0 +1,12 @@
+package com.ujutechnology.api8.api.dto.rentHouseLoan;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class RentHouseResult {
+    RentHouseOuterWrapperResult result;
+}
