--- src/main/java/com/ujutechnology/api8/biz/dto/creditLoan/CreditLoanResult.java
@@ -1,5 +1,0 @@
-package com.ujutechnology.api8.biz.dto.creditLoan;
-
-public class CreditLoanResult {
-    CreditLoanOuterWrapperResult result;
-}
