--- src/main/java/com/ujutechnology/api8/biz/dto/deposit/DepositResult.java
@@ -1,5 +1,0 @@
-package com.ujutechnology.api8.biz.dto.deposit;
-
-public class DepositResult {
-    DepositOuterWrapperResult result;
-}
