--- src/main/java/com/ujutechnology/api8/biz/dto/mortgageLoan/MortgageResult.java
@@ -1,5 +1,0 @@
-package com.ujutechnology.api8.biz.dto.mortgageLoan;
-
-public class MortgageResult {
-    MortgageOuterWrapperResult result;
-}
