--- src/main/java/com/ujutechnology/api8/biz/dto/rentHouseLoan/RentHouseResult.java
@@ -1,5 +1,0 @@
-package com.ujutechnology.api8.biz.dto.rentHouseLoan;
-
-public class RentHouseResult {
-    RentHouseOuterWrapperResult result;
-}
